feat(admin): 文章模块的文章、分类和标签基本对接

+++ apps/backend/src/blog/service/class.service.ts
@@ -1,18 +1,24 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
-import { ClassModel } from '../../models';
+import { ArticleClassModel, ClassModel } from '../../models';
 import { Res } from '../../response';
 import { PageDto } from '../dto/page';
 import { getPageOffset, getUpdateData } from '../../utils';
 import { AddDto } from '../dto/class/add.dto';
 import { UpdateDto } from '../dto/class/update.dto';
 import dayjs from 'dayjs';
+import { ListDto } from '../dto/class/list.dto';
+import { Attributes, FindAndCountOptions } from 'sequelize/types/model';
+import { Op } from 'sequelize';
+import { isString } from '../../utils/is';
 
 @Injectable()
 export class ClassService {
   constructor(
     @InjectModel(ClassModel)
     private classModel: typeof ClassModel,
+    @InjectModel(ArticleClassModel)
+    private articleClassModel: typeof ArticleClassModel,
   ) {}
 
   //查询所有
@@ -26,14 +32,30 @@ export class ClassService {
   }
 
   //分页查询
-  async listByPage(data: PageDto) {
+  async listByPage(data: ListDto) {
     try {
-      const offset = getPageOffset(data.page, data.limit);
-      const result = await this.classModel.findAndCountAll({
-        limit: data.limit,
-        offset: offset,
-      });
-      return Res.OKWithPage(result.rows, data.limit, data.page, result.count);
+      const offset = getPageOffset(data.page, data.pageSize);
+      const options: Omit<FindAndCountOptions<Attributes<ClassModel>>, 'group'> = {
+        limit: data.pageSize,
+        offset,
+      };
+      const whereOptions = {};
+      //如果有分类名，模糊匹配
+      if (data.name) {
+        whereOptions['name'] = {
+          [Op.like]: `%${data.name}%`,
+        };
+      }
+      //创建时间范围内
+      if (isString(data.startTime) && isString(data.endTime)) {
+        whereOptions['createdAt'] = {
+          [Op.gte]: new Date(data.startTime),
+          [Op.lte]: new Date(data.endTime),
+        };
+      }
+      options.where = whereOptions;
+      const result = await this.classModel.findAndCountAll(options);
+      return Res.OKWithPage(result.rows, data.pageSize, data.page, result.count);
     } catch (e) {
       return Res.Error(e.message);
     }
@@ -120,15 +142,27 @@ export class ClassService {
     }
   }
 
-  async batchDelete(ids: string) {
+  //删除数据
+  async deleteData(ids: string) {
     try {
       const idList = ids.split(',').map((id) => parseInt(id));
+      const record = await this.articleClassModel.findOne({
+        where: {
+          class_id: {
+            [Op.in]: idList,
+          },
+        },
+      });
+      if (record) {
+        //分类已经被使用
+        return Res.Error('该分类已经被文章引用');
+      }
       const count = await this.classModel.destroy({
         where: {
           id: idList,
         },
       });
-      return Res.OKWithData(count, count > 0 ? '删除成功' : '有id不存在');
+      return Res.OKWithData(count, '删除成功');
     } catch (e) {
       return Res.Error(e.message);
     }
